more tidy-up
